requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
jwt==1.3.1
Markdown==3.5.2
pycparser==2.21
pytz==2024.1
sqlparse==0.4.4
